% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Candidate.R
\name{createCandidate}
\alias{createCandidate}
\title{A wrapper function for \code{Candidate}'s constructor,
which allows consumers to ignore the automatic generation
of the delegatesNeeded value.}
\usage{
createCandidate(name, delegatesWon, party)
}
\arguments{
\item{name}{The candidate's name.}

\item{delegatesWon}{The number of delegates won by the candidate.}

\item{party}{The candidate's party.}
}
\value{
A new Candidate object, initialized according to the provided values.
}
\description{
A wrapper function for \code{Candidate}'s constructor,
which allows consumers to ignore the automatic generation
of the delegatesNeeded value.
}

