% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Candidate.R
\name{delegates.needed}
\alias{delegates.needed}
\title{Computes the number of delegates needed by a candidate
of the specified party.}
\usage{
delegates.needed(party)
}
\arguments{
\item{party}{The party to check.}
}
\value{
The number of delegates necessary for a member of
	the specified party to secure a nomination.
}
\description{
Computes the number of delegates needed by a candidate
of the specified party.
}

