% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_Candidate.R
\docType{methods}
\name{initialize,Candidate-method}
\alias{initialize,Candidate-method}
\title{Initializes a Candidate object according to the provided values,
validating that the candidate's party is either 'Republican' or 'Democratic'.
Used instead of a validation function because it accomplishes the tasks
of initialization and validation simultaneously.}
\usage{
\S4method{initialize}{Candidate}(.Object, name, delegatesWon, party, ...)
}
\arguments{
\item{.Object}{Candidate. The candidate to initialize.}
}
\value{
A new Candidate with the provided values.
}
\description{
Automatically generates the value of delegatesNeeded.
}

