% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_Candidate.R
\docType{methods}
\name{print,Candidate-method}
\alias{print,Candidate-method}
\title{Prints a Candidate object, by printing all its fields,
in the order: name, party, delegatesWon, delegatesNeeded.}
\usage{
\S4method{print}{Candidate}(x, ...)
}
\arguments{
\item{x}{Candidate. The Candidate to print.}
}
\description{
Prints a Candidate object, by printing all its fields,
in the order: name, party, delegatesWon, delegatesNeeded.
}

