% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propNeeded_Candidate.R
\docType{methods}
\name{propNeeded,Candidate-method}
\alias{propNeeded,Candidate-method}
\title{Computes the proportion of delegates a candidate still must
win to secure a nomination.}
\usage{
\S4method{propNeeded}{Candidate}(object)
}
\arguments{
\item{object}{Candidate. The candidate to calculate against.}
}
\value{
The proportion of delegates the provided candidate must win.
}
\description{
Computes the proportion of delegates a candidate still must
win to secure a nomination.
}

