% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propNeeded_Candidate.R
\name{propNeeded}
\alias{propNeeded}
\title{Computes the proportion of whatever the provided object measures
that is still necessary.}
\usage{
propNeeded(object)
}
\arguments{
\item{object.}{The object to calculate against.}
}
\description{
Computes the proportion of whatever the provided object measures
that is still necessary.
}

