% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_Candidate.R
\docType{methods}
\name{show,Candidate-method}
\alias{show,Candidate-method}
\title{Applies the method `show` to Candidate object.
The equivalent of calling `print`.}
\usage{
\S4method{show}{Candidate}(object)
}
\arguments{
\item{object}{Candidate. The candidate to show.}
}
\description{
Applies the method `show` to Candidate object.
The equivalent of calling `print`.
}
\seealso{
print.Candidate
}

